Added more styles - Added exceptions for delete and inputs

+++ public/js/comment.js
@@ -4,13 +4,17 @@ const newCommentFormHandler = async (event) => {
 
   const comment = document.querySelector('.input-comment').value.trim();
   const post_id = document.querySelector('.form-add-comment').getAttribute('data-post-id');
-  const user_id = document.querySelector('.form-add-comment').getAttribute('data-user-id');
+
+  if (!comment) {
+    alert("Your comment cannot be empty - Please add a comment")
+    return;
+  }
 
   const response = await fetch('/api/comments/', {
     method: 'POST',
     body: JSON.stringify({
       comment: comment,
-      post_id: post_id, 
+      post_id: post_id,
     }),
     headers: {
       'content-type': 'application/json',
@@ -34,15 +38,22 @@ const delCommentBtnEl = document.querySelector('.button-delete-comment');
 const commentId = delCommentBtnEl.getAttribute('data-id');
 
 const deleteComment = async () => {
-  const response = await fetch(`/api/comments/${commentId}`, {
-    method: 'DELETE',
-  });
+  if (confirm('ARE YOU SURE YOU WANT TO DELETE COMMENT?') == true) {
+    const response = await fetch(`/api/comments/${commentId}`, {
+      method: 'DELETE',
+    });
 
-  if (response.ok) {
-    window.location.replace(window.location.pathname)
-  } else {
-    alert('Failed to delete comment');
-  };
+    if (response.ok) {
+      window.location.replace(window.location.pathname)
+    } else {
+      alert('Failed to delete comment');
+    };
+  }
+  else {
+    return;
+  }
+
+
 };
 
 delCommentBtnEl.addEventListener('click', deleteComment);
